requirements.txt
numpy==1.23.1 
matplotlib==3.7.4 
opencv-python 
pillow==10.2.0
pytorch-lightning==2.1.3 
hydra-core==1.3.2 
open3d==0.18.0 
trimesh==4.0.9
scikit-image==0.21.0 
scikit-learn==1.3.2 
scipy==1.10.1 
sk-video==1.1.10 
pandas
tqdm==4.65.0 
rich==13.7.0 
easydict==1.11 
omegaconf==2.3.0 
typing-extensions==4.4.0 
pyyaml==6.0.1 
timm==0.9.12
pytorch-msssim==1.0.0 
joblib==1.3.2 
albumentations==0.5.2 
kornia==0.5.0 
webdataset==0.2.86  
pybullet==3.2.6
tensorboard==2.14.0
rtree==1.2.0
lpips==0.1.4
prettytable==3.9.0
chumpy==0.70
wandb==0.16.2
pymeshlab
objaverse==0.1.7
git+https://github.com/facebookresearch/pytorch3d.git   # version 0.7.5
-e submodules/Inpaint_Anything/segment_anything
